Made navigation bar a reuable component in React
